Integrate parser code into gptokeyb.cpp
It's not really necessary for this to be in a separate file.
